Made edits to game logic and UI messaging
Made changes to checkTie and onClickBoard functions, adding UI messaging
so that users can see whose turn it is to make a move in tic tac toe
game. Set display to none for change password and log-out elements for
when page first renders. Created GET request to and added button
that when clicked will retrieve the total number of games played
for a logged-in user.